+++ embedded/common/cross_door.rs
@@ -5,12 +5,12 @@
 //! Cross door behaviors.
 
 use alloc::sync::Arc;
-
+// use ariel_os::time::Timer;
 use behaviortree::prelude::*;
 use spin::Mutex;
 
 fn sleep_ms(millisecs: u64) {
-	//	thread::sleep(Duration::from_millis(millisecs));
+	// Timer::after_millis(millisecs).await;
 }
 
 /// `CrossDoor` behavior interface